Cast return of strtol to unsigned int
strtol returns long, which is compared against unsigned int.  On 32-bit
hosts, it leads to

gcc/gentarget-def.c: In function void def_target_insn(const char*, const char*):
gcc/gentarget-def.c:88:34: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]

This patch casts return of strtol to unsigned int to avoid the error.

	* gentarget-def.c (def_target_insn): Cast return of strtol to
	unsigned int.

From-SVN: r224993

+++ gcc/gentarget-def.c
@@ -85,7 +85,7 @@ def_target_insn (const char *name, const char *prototype)
 	   That doesn't contribute to the suffix, so skip ahead and
 	   process the following character.  */
 	char *endptr;
-	if (strtol (p + 1, &endptr, 10) != opno
+	if ((unsigned int) strtol (p + 1, &endptr, 10) != opno
 	    || (*endptr != ',' && *endptr != ')'))
 	  {
 	    error ("invalid prototype for '%s'", name);
